[refactor] Moved module logic into module server
